requirements.txt
numpy~=1.26.4
scipy~=1.13.1
scikit-learn~=1.5.1
pandas~=2.2.2
numba~=0.60.0
setuptools~=72.1.0
tqdm~=4.66.5
obspy~=1.4.1
PyYAML~=6.0.2
torch~=2.4.0
packaging~=24.1
matplotlib~=3.9.1.post1
seaborn~=0.13.2
h5py~=3.11.0
requests~=2.32.3
lxml~=5.2.2
Bottleneck~=1.4.0
ipython~=8.18.1
pyproj~=3.6.1
pytorch_lightning~=2.4.0
nest_asyncio~=1.6.0
tensorboardX
tensorboard
seisbench